Move quantized mesh loader to terrain module
