--- modules/quantized-mesh/README.md
@@ -1,9 +1,0 @@
-# @loaders.gl/quantized-mesh
-
-[loaders.gl](https://loaders.gl/docs) is a collection of framework independent 3D and geospatial parsers and encoders.
-
-This module reconstructs mesh surfaces from the [quantized mesh](https://github.com/CesiumGS/quantized-mesh) format.
-
-For documentation please visit the [website](https://loaders.gl).
-
-`@loaders.gl/quantized-mesh` uses [`quantized-mesh-decoder`](https://github.com/heremaps/quantized-mesh-decoder) from HERE under the MIT license to decode quantized mesh.
