--- modules/quantized-mesh/docs/README.md
@@ -1,14 +1,0 @@
-# Overview
-
-The `@loaders.gl/quantized-mesh` module reconstructs mesh surfaces from the [quantized mesh](https://github.com/CesiumGS/quantized-mesh) format.
-
-## Installation
-
-```bash
-npm install @loaders.gl/quantized-mesh
-npm install @loaders.gl/core
-```
-
-## Attribution
-
-`@loaders.gl/quantized-mesh` uses [`quantized-mesh-decoder`](https://github.com/heremaps/quantized-mesh-decoder) from HERE under the MIT license to decode quantized mesh.
