--- modules/quantized-mesh/package.json
@@ -1,39 +1,0 @@
-{
-  "name": "@loaders.gl/quantized-mesh",
-  "version": "2.1.0-beta.2",
-  "description": "Framework-independent loader for quantized mesh",
-  "license": "MIT",
-  "publishConfig": {
-    "access": "public"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/visgl/loaders.gl"
-  },
-  "keywords": [
-    "webgl",
-    "loader",
-    "3d",
-    "mesh",
-    "point cloud",
-    "OBJ"
-  ],
-  "main": "dist/es5/index.js",
-  "module": "dist/esm/index.js",
-  "esnext": "dist/es6/index.js",
-  "sideEffects": false,
-  "files": [
-    "src",
-    "dist",
-    "README.md"
-  ],
-  "scripts": {
-    "pre-build": "npm run build-worker && npm run build-bundle && npm run build-bundle -- --env.dev",
-    "build-bundle": "webpack --display=minimal --config ../../scripts/bundle.config.js",
-    "build-worker": "webpack --entry ./src/quantized-mesh-loader.worker.js --output ./dist/quantized-mesh-loader.worker.js --config ../../scripts/worker-webpack-config.js"
-  },
-  "dependencies": {
-    "@babel/runtime": "^7.3.1",
-    "@loaders.gl/loader-utils": "^2.1.1"
-  }
-}
