--- modules/quantized-mesh/src/index.js
@@ -1,1 +1,0 @@
-export {QuantizedMeshLoader, QuantizedMeshWorkerLoader} from './quantized-mesh-loader';
