--- modules/quantized-mesh/test/data/README.md
@@ -1,2 +1,0 @@
-`tile-with-extensions.terrain` is from [Cesium World Terrain from Cesium
-ion](https://cesiumjs.org/Cesium/Build/Apps/Sandcastle/index.html?src=Terrain.html)
