--- modules/quantized-mesh/test/index.js
@@ -1,1 +1,0 @@
-import './quantized-mesh-loader.spec';
