+++ modules/terrain/src/index.js
@@ -1,1 +1,2 @@
 export {TerrainLoader, TerrainWorkerLoader} from './terrain-loader';
+export {QuantizedMeshLoader, QuantizedMeshWorkerLoader} from './quantized-mesh-loader';
