+++ modules/terrain/src/lib/decode-quantized-mesh.js
