+++ modules/terrain/src/quantized-mesh-loader.worker.js
