+++ modules/terrain/test/data/README.md
@@ -1,3 +1,6 @@
 mapbox.png is from Mapbox's [Terrain RGB service](https://docs.mapbox.com/help/troubleshooting/access-elevation-data/#mapbox-terrain-rgb)
 
 terrarium.png is from [Mapzen Terrain Tiles](https://registry.opendata.aws/terrain-tiles/).
+
+`tile-with-extensions.terrain` is from [Cesium World Terrain from Cesium
+ion](https://cesiumjs.org/Cesium/Build/Apps/Sandcastle/index.html?src=Terrain.html)
