+++ modules/terrain/test/data/tile-with-extensions.terrain
