+++ modules/terrain/test/index.js
@@ -1,1 +1,2 @@
 import './terrain-loader.spec';
+import './quantized-mesh-loader.spec';
